Preparation for the segmented bars

+++ Assets/Scripts/MidSummonController.cs
@@ -19,9 +19,9 @@ public class MidSummonController : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        curantHealth = summonStats.maxHealth;
+        curantHealth = (int)summonStats.maxHealth;
         dgmOnCountry = summonStats.dmgOnCountry;
-        damege = summonStats.damegeDelt;
+        damege = (int)summonStats.damegeDelt;
         death = false;
         takeDamage = false;
         animator = this.gameObject.GetComponent<Animator>();
@@ -43,8 +43,8 @@ public class MidSummonController : MonoBehaviour
     {
         if (summonStats.updateBool == true)
         {
-            damege = summonStats.damegeDelt;
-            curantHealth += summonStats.maxHealth - summonStats.healthMax;
+            damege = (int)summonStats.damegeDelt;
+            curantHealth += (int)summonStats.maxHealth - summonStats.healthMax;
             summonStats.updateBool = false;
         }
         if (curantHealth <= 0)
